+++ flake.nix
@@ -72,28 +72,13 @@
           inherit system;
         };
 
-        packages.book =
-          let
-            inherit (pkgs) lib;
-            root = ./.;
-            rootPrefix = toString root;
-            cleanedSrc = lib.cleanSourceWith {
-              src = root;
-              filter = path: type:
-                let
-                  relativePath = lib.removePrefix rootPrefix path;
-                in
-                lib.any (prefix: lib.hasPrefix prefix relativePath) [
-                  "/docs" # Build the docs directory
-                  "/examples" # But also include examples as we cross-reference them
-                  "/README.md"
-                  "/CHANGELOG.md"
-                ];
-            };
-          in
-          pkgs.runCommand "crane-book" { } ''
-            ${pkgs.mdbook}/bin/mdbook build --dest-dir $out ${cleanedSrc}/docs
-          '';
+        # To override do: lib.overrideScope' (self: super: { ... });
+        lib = mkLib pkgs;
+
+        packages = import ./pkgs {
+          inherit pkgs;
+          myLib = lib;
+        };
 
         checks =
           let
@@ -107,9 +92,6 @@
             myLib = mkLib pkgsChecks;
             myPkgs = packages;
           };
-
-        # To override do: lib.overrideScope' (self: super: { ... });
-        lib = mkLib pkgs;
       in
       {
         inherit checks lib packages;