+++ lib/downloadCargoPackageFromGit.nix
@@ -1,7 +1,7 @@
 { cargo
+, craneUtils
 , jq
 , lib
-, remarshal
 , runCommandLocal
 }:
 
@@ -21,8 +21,8 @@ let
   deps = {
     nativeBuildInputs = [
       cargo
+      craneUtils
       jq
-      remarshal
     ];
   };
 in
@@ -30,8 +30,9 @@ runCommandLocal "cargo-git" deps ''
   mkdir -p $out
   existing_crates=()
   while read -r cargoToml; do
-    local crate=$(toml2json <"$cargoToml" | \
-      jq -r 'select(.package != null) | .package | "\(.name)-\(.version)"'
+    local crate=$(
+      cargo metadata --format-version 1 --no-deps --manifest-path "$cargoToml" |
+      jq -r '.packages[] | select(.manifest_path == "'"$cargoToml"'") | "\(.name)-\(.version)"'
     )
 
     if [ -n "$crate" ]; then
@@ -44,6 +45,9 @@ runCommandLocal "cargo-git" deps ''
       cp -r "$(dirname "$cargoToml")" "$dest"
       chmod +w "$dest"
       echo '{"files":{}, "package":null}' > "$dest/.cargo-checksum.json"
+
+      crane-resolve-workspace-inheritance "$cargoToml" > "$dest/Cargo.toml.resolved" &&
+        mv "$dest/Cargo.toml"{.resolved,}
 
       existing_crates+=("$crate")
     fi