+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 * Previously compiled build scripts now maintain their executable bit when
   inherited
+* Workspace inheritance in git dependencies is now correctly handled
 
 ## [0.11.0] - 2022-12-26